# docs/openapi/leadflow-api.yaml
openapi: 3.0.3
info:
  title: LeadFlow API
  version: 1.0.0
  description: API REST do CRM B2B LeadFlow.
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
paths:
...
  /leads/{id}:
    get:
      summary: Obter Lead
      parameters: [{ name: id, in: path, required: true, schema: { type: integer, format: int64 } }]
      responses: { '200': { description: Lead } }
    put:
      summary: Editar Lead
      parameters: [{ name: id, in: path, required: true, schema: { type: integer, format: int64 } }]
      responses: { '200': { description: Lead atualizado } }
  /leads/{id}/stage:
    patch:
      summary: Alterar etapa e registrar histórico
      parameters: [{ name: id, in: path, required: true, schema: { type: integer, format: int64 } }]
      responses: { '200': { description: Etapa atualizada } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT